+++ tensorpc/dock/client.py
@@ -74,13 +74,6 @@ class MasterMeta:
         self.grpc_url = grpc_url
         self.http_url = url
         self.fwd_http_port = fwd_http_port
-        lsp_port = os.getenv(constants.TENSORPC_FLOW_APP_LANG_SERVER_PORT,
-                             None)
-        self.lsp_port = int(lsp_port) if lsp_port is not None else None
-        lsp_fwd_port = os.getenv(
-            constants.TENSORPC_FLOW_APP_LANG_SERVER_FWD_PORT, None)
-        self.lsp_fwd_port = int(
-            lsp_fwd_port) if lsp_fwd_port is not None else None
 
         self.is_grpc_valid = grpc_url != ""
         self.is_http_valid = self.http_url != ""