+++ tensorpc/dock/components/three/base.py
@@ -219,15 +219,17 @@ class DataPortalProps(ContainerBaseProps):
     comps: List[Component] = dataclasses.field(default_factory=list)
 
 class DataPortal(ThreeContainerBase[DataPortalProps, ThreeComponentType]):
-    def __init__(self, comps: List[Component], children: Optional[ThreeLayoutType] = None) -> None:
+    def __init__(self, source: Component, children: Optional[ThreeLayoutType] = None) -> None:
         if children is not None and isinstance(children, Sequence):
             children = {str(i): v for i, v in enumerate(children)}
         allowed_comp_types = {UIType.DataModel, UIType.ThreeURILoaderContext, UIType.ThreeCubeCamera}
-        for comp in comps:
+        
+        sources = [source]
+        for comp in sources:
             assert comp._flow_comp_type in allowed_comp_types, "DataPortal only support DataModel and resource loaders."
-        assert len(comps) > 0, "comps must have at least one component"
+        assert len(sources) == 1, "DataPortal only support one source."
         super().__init__(UIType.DataPortal, DataPortalProps, children, allowed_events=[])
-        self.prop(comps=comps)
+        self.prop(comps=sources)
 
     @property
     def prop(self):